+++ main.go
@@ -13,19 +13,42 @@ func main() {
 
 	_, err := os.Stat(*cfgPath)
 	if os.IsNotExist(err) {
-		log.Fatalf("config file %s does not exist", *cfgPath)
+		log.Fatalf("config file %s does not exist, creating from template", *cfgPath)
+
+		var file *os.File
+		file, err = os.Create(*cfgPath)
+		if err != nil {
+			log.Fatalf("failed to create temaplate '%s' with error %v", *cfgPath, err)
+			os.Exit(1)
+		}
+		defer file.Close()
+
+		var templateJSONBytes []byte
+		templateJSONBytes, err = monkebot.ConfigTemplateJSON()
+		if err != nil {
+			log.Fatalf("failed to create temaplate '%s' with error %v", *cfgPath, err)
+			os.Exit(1)
+		}
+
+		_, err = file.Write(templateJSONBytes)
+		if err != nil {
+			log.Fatalf("failed to create temaplate '%s' with error %v", *cfgPath, err)
+			os.Exit(1)
+		}
 	}
 
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	cfg, err := monkebot.LoadConfigFromFile(*cfgPath)
+	var cfg *monkebot.Config
+	cfg, err = monkebot.LoadConfigFromFile(*cfgPath)
 	if err != nil {
 		log.Fatal(err)
 	}
 
-	mb, err := monkebot.NewMonkebot(*cfg)
+	var mb *monkebot.Monkebot
+	mb, err = monkebot.NewMonkebot(*cfg)
 	if err != nil {
 		log.Fatal(err)
 	}