+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # Changes
-- Binary release
+- Show an error when the config file is missing a field
+- Save a template as config.json if the file does not exist
+- Test to ensure templates have all Config fields